# Optimize array iterators