Their performance had regressed since 2.12. Now they are consistently
faster than the 2.12 implementation.

Fixes https://github.com/scala/bug/issues/11118

Before:

```
[info] Benchmark                              (valueCount)   Mode  Cnt           Score          Error  Units
[info] ArrayIteratorBenchmarks.arrayIterator             0  thrpt    5  1062464845.202 ± 24904701.044  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator           100  thrpt    5     1478433.436 ±    39484.575  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator          1000  thrpt    5      148765.505 ±     2190.330  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator         10000  thrpt    5       14901.314 ±      364.297  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator        100000  thrpt    5        1459.279 ±       72.760  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator       1000000  thrpt    5         144.065 ±        4.327  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator      10000000  thrpt    5          14.385 ±        0.800  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator     100000000  thrpt    5           1.452 ±        0.052  ops/s
```

After:

```
[info] Benchmark                              (valueCount)   Mode  Cnt           Score          Error  Units
[info] ArrayIteratorBenchmarks.arrayIterator             0  thrpt    5  1053249647.465 ± 33697918.043  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator           100  thrpt    5     1589050.782 ±    64432.883  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator          1000  thrpt    5      157945.945 ±     3639.108  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator         10000  thrpt    5       17558.844 ±      719.517  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator        100000  thrpt    5        1837.055 ±       70.482  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator       1000000  thrpt    5         185.288 ±        9.067  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator      10000000  thrpt    5          18.194 ±        0.893  ops/s
[info] ArrayIteratorBenchmarks.arrayIterator     100000000  thrpt    5           1.831 ±        0.042  ops/s
```