@szeiger 2018-09-11
In case someone is wondering about the lack of `length == 0` optimization (returning `Iterator.empty`): I tried it and it was significantly slower for the length == 0 case (no measurable difference in other cases).