@viktorklang 2018-09-11
@szeiger If it was only slower for the 0-case, wouldn't that indicate that the empty iterator needs to be looked at in terms of performance?