@NthPortal 2018-09-12
@viktorklang either that or it's getting deoptimised because of being megamorphic?