@viktorklang 2018-09-12
@NthPortal If that's the case then I misunderstood—because I thought the bench where *only* benching 0-length arrays, in which case it would always return the empty iterator, hence be monomorphic even if it needs to call through invokeinterface.